pullBackendProject and initializeBackendProject renames
